BROKEN: failed attempt to make votes persistent
Vote options were previously handled in the non-persistent hashmaps
for mock data. Currently, spring-data-mongo does not handle the
vote-options correctly. I tried to make a transactional method that
retrieves the poll, changes it using the previously working logic,
and stores it using pollrepository.save(). It works except for the save,
that gives an uncomprehensible error in the console. :(