Redesigned REST API, renamed resources in frontend
Changed the paths, both in API and in frontend to follow convention.
For REST, the words in the path should be plural nouns denoting the
resource that is intuitively found at this path.
For frontend, it can be the same, but in some cases it should be an
action such as register or login